+++ libs/bash/git/install.sh
@@ -5,6 +5,7 @@ set -e
 # usage ./libs/bash/install.sh (optional) branch_name
 # example ./libs/bash/install.sh fix/ICON-39280/connect_preprod_module_revert_to_0_36_7_due_to_kms_permissions
 
+# Version: 0.6.0  - 2024-05-24 - UPDATED Git hook symlink creation - David J Eddy
 # Version: 0.5.11 - 2024-05-06 - ADD Git feature checks - David J Eddy
 # Version: 0.5.10 - 2024-04-22
 # Version: 0.5.8  - 2024-03-19
@@ -23,7 +24,7 @@ fi
 if [[ ! $WORKSPACE ]]
 then
     declare WORKSPACE
-    WORKSPACE="$(git rev-parse --show-toplevel)"
+    WORKSPACE=$(git rev-parse --show-toplevel)
     export WORKSPACE
 fi
 printf "INFO: WORKSPACE %s\n" "${WORKSPACE}"
@@ -46,6 +47,7 @@ printf "INFO: Removing existing .tmp if exists.\n"
 rm -rf "$WORKSPACE/.tmp" || exit 1
 printf "INFO: Clone toolchain-management project locally into %s/.tmp\n" "$(pwd)"
 git clone --quiet git@gitlab.test.igdcs.com:cicd/terraform/tools/toolchain-management.git "$WORKSPACE/.tmp/toolchain-management"
+
 # Even if main, checkout anyways
 cd "$WORKSPACE/.tmp/toolchain-management"
 git checkout "$WL_GC_TOOLCHAIN_BRANCH" --force
@@ -54,18 +56,20 @@ cd "$WORKSPACE" || exit 1
 printf "INFO: Execute toolchain-management installer...\n"
 "$WORKSPACE/.tmp/toolchain-management/libs/bash/install.sh" "$@"
 
-printf "INFO: Installing Git pre-commit hooks.\n"
-rm -rf "$WORKSPACE/.git/hooks/pre-commit" || true
-ln -sfn "$WORKSPACE/.tmp/toolchain-management/libs/bash/git/pre-commit.sh" "$WORKSPACE/.git/hooks/pre-commit"
-
-printf "INFO: Installing Git pre-push hooks.\n"
-rm -rf "$WORKSPACE/.git/hooks/pre-push" || true
-ln -sfn "$WORKSPACE/.tmp/toolchain-management/libs/bash/git/pre-push.sh" "$WORKSPACE/.git/hooks/pre-push"
+# create symlink for each hook found
+declare GIT_HOOKS
+GIT_HOOKS=$(find "$WORKSPACE/.tmp/toolchain-management/libs/bash/git/hooks" -maxdepth 1 -type f | sed "s/.*\///")
+for HOOK in $GIT_HOOKS
+do
+    printf "INFO: Installing Git %s hook.\n" "${HOOK}"
+    rm -rf "$WORKSPACE/.git/hooks/${HOOK}" || true
+    ln -sfn "$WORKSPACE/.tmp/toolchain-management/libs/bash/git/hooks/${HOOK}" "$WORKSPACE/.git/hooks/${HOOK}"
+done
 
 # Git features
 if [[ -f .gitattributes ]]
 then
-    printf "INFO: Configure git LFS.\n"
+    printf "INFO: Configure Git LFS.\n"
     git lfs track "*.iso"
     git lfs track "*.zip"
     git lfs track "*.gz"